erp-backend/app/modules/raffles/pyproject.toml
[project]
name = "raffle-backend"
version = "1.0.0"
description = "Service for Jungle One's raffles . Welcome tho the Jungle!"
authors = [
    {name = "Jungle One S.A.S.",email = "jungleone.team@gmail.com"}
]
license = {text = "MIT"}
readme = "../README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "gunicorn (>=22.0.0,<23.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    # CAMBIO 1: Se modifica 'passlib' para usar un formato más explícito.
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    # CAMBIO 2: Se añade y fija la versión de 'bcrypt' para asegurar la compatibilidad.
    "bcrypt==3.2.2",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiohttp (>=3.11.15)",
    "pillow (>=11.2.1,<12.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]

[tool.poetry]
package-mode = false  # Explicitly disable packaging

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:main"  # Makes `poetry run start` work